Remove database/json/models in favor of JsonTable
Instead of having database-specific models for the JsonDatabase, we'll
have a single model with an array as the root element (versus having a
dedicated model with a typed array as the root). With this model, we cut
down on database-specific modeling, and instead, can re-use the existing
data classes from the server system. This setup also allows for more
abstracted table methods like storage and retrieval.

Test: ./gradlew clean test