Change JsonTable storage from List to Map
Changing the storage method of JsonTable from a list to a map allows for
primary key lookup where the primary key can be defined arbitrarily by
the model. This enables much faster lookup, deletion, and replacement at
the cost of memory. In our case, this tradeoff is acceptable as our
memory footprint will still be O(n).

Test: ./gradlew clean test