Rename Result->LegacyResult, add new Result
Legacy Result.httpStatusCode is a nightmare to deal with as it gives
too much freedom in terms of `when` handling the return values. There
is no static checking against HttpStatusCodes to verify all possible
statuses are accounted for. Hence, the new Result data class is defined
as an Enum which leverages the Kotlin compiler to require `when`
statements to handle all cases when handling the enum. A helper
interface `StatusEnum` is also defined to equate each enum value to a
single HttpStatusCode.

This is the initial change to introduce the Result v2 data class.
Follow-up commits will replace usage of LegacyResult. Eventually,
LegacyResult will be completely removed.

Test: ./gradlew clean test