Remove StatusEnum constraint for Result
Internal methods, especially those of the controller, should be ignorant
to the HTTP status that should occur as a result of its own method
calls. The status enum in and of itself should provide enough context
for the routing handler to justify one status code over another. This
also frees the router from needing to context switch whether or not it
should trust the Result's http status, or provide its own.

Test: ./gradlew clean test