Remove references to User.username
Instead of assuming the uniqueness of User.username to identify users,
methods should reference the new UserId. This data structure will be
used as the single source of uniqueness for users allowing for the user
body structure to be modified without needing major refactors. Methods
that need a reference to a user should use UserId and fetch and update
their own copy of the User instead of passing the entire User structure
across components.

Test: ./gradlew clean test